% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Make Shiny Application to Executable File}
\usage{
pack(app_name = "myapp", electron_settings = list())
}
\arguments{
\item{app_name}{Name of your application. Default will be "myapp".}

\item{electron_settings}{a list including package.json settings. Including product_name, app_version, app_description, author_name, author_email, repository_url}
}
\description{
This Function make your shiny app to an executable file
Go to your project directory(Including app.R), and run this function.
}
\examples{
\dontrun{
pack(
  app_name = "myapp",
  electron_settings = list(
    c("product_name_template", "My Own Product Name"),
    c("app_description_template", "App Description"),
    c("author_name_template", "Author Name"),
    c("author_email_template", "Author E-mail"),
    c("repository_url_template", "Repository URL")
  )
)
}
}
\author{
Changwoo Lim
}
